helpers/upload_release/requirements.txt
fabric3
python-gitlab
semver

